requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
google-generativeai==0.7.2
python-dotenv==1.0.0
